Move AndroidHelper to components folder and remove namespace.
